+++ C#/MustDo.Solution/MustDo.Domain/Entities/Categoria.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace MustDo.Domain.Entities
+{
+	public class Categoria
+	{
+		public Guid CategoriaId { get; set; } = Guid.NewGuid();
+		public string Nome { get; set; }
+	}
+}