+++ C#/MustDo.Solution/MustDo.Domain/Interfaces/Repositories/ICategoriaRepository.cs
@@ -1,0 +1,9 @@
+﻿using MustDo.Domain.Entities;
+
+namespace MustDo.Domain.Interfaces.Repositories
+{
+	public interface ICategoriaRepository : IRepositoryBase<Categoria>
+	{
+
+	}
+}