+++ C#/MustDo.Solution/MustDo.Domain/Interfaces/Services/ICategoriaService.cs
@@ -1,0 +1,9 @@
+﻿using MustDo.Domain.Entities;
+
+namespace MustDo.Domain.Interfaces.Services
+{
+	public interface ICategoriaService : IServiceBase<Categoria>
+	{
+
+	}
+}