+++ C#/MustDo.Solution/MustDo.Infra.Data/Contexts/MustDo4EntitiesDb.cs
@@ -1,0 +1,43 @@
+﻿using MustDo.Domain.Entities;
+using MustDo.Infra.Data.Mappings;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Data.Entity;
+using System.Data.Entity.ModelConfiguration.Conventions;
+
+namespace MustDo.Infra.Data.Contexts
+{
+	public class MustDo4EntitiesDb : DbContext
+	{
+		public DbSet<Categoria> Categorias { get; set; }
+		public DbSet<Comentario> Comentarios { get; set; }
+		public DbSet<Tarefa> Tarefas { get; set; }
+
+		protected override void OnModelCreating(DbModelBuilder modelBuilder)
+		{
+			base.OnModelCreating(modelBuilder);
+
+			// Conventions
+			modelBuilder.Conventions.Remove<PluralizingTableNameConvention>();
+			modelBuilder.Conventions.Remove<OneToManyCascadeDeleteConvention>();
+			modelBuilder.Conventions.Remove<ManyToManyCascadeDeleteConvention>();
+
+			// General Custom Context Properties
+			modelBuilder.Properties()
+			    .Configure(p => p.HasDatabaseGeneratedOption
+			    (DatabaseGeneratedOption.None));
+
+			modelBuilder.Properties<string>()
+			    .Configure(p => p.HasColumnType("varchar"));
+
+
+			//AddConfigMap(modelBuilder);
+		}
+
+		private void AddConfigMap(DbModelBuilder modelBuilder)
+		{
+			modelBuilder.Configurations.Add(new CategoriaMap());
+			modelBuilder.Configurations.Add(new ComentarioMap());
+			modelBuilder.Configurations.Add(new TarefaMap());
+		}
+	}
+}