+++ C#/MustDo.Solution/MustDo.Infra.Data/Mappings/CategoriaMap.cs
@@ -1,0 +1,17 @@
+﻿using MustDo.Domain.Entities;
+using System.Data.Entity.ModelConfiguration;
+
+namespace MustDo.Infra.Data.Mappings
+{
+	public class CategoriaMap : EntityTypeConfiguration<Categoria>
+	{
+		public CategoriaMap()
+		{
+			ToTable("Categorias");
+
+			HasKey(p => p.CategoriaId);
+
+			Property(p => p.Nome).IsRequired().HasMaxLength(50);
+		}
+	}
+}