+++ C#/MustDo.Solution/MustDo.Infra.Data/Mappings/TarefaMap.cs
@@ -1,0 +1,18 @@
+﻿using MustDo.Domain.Entities;
+using System.Data.Entity.ModelConfiguration;
+
+namespace MustDo.Infra.Data.Mappings
+{
+	public class TarefaMap : EntityTypeConfiguration<Tarefa>
+	{
+		public TarefaMap()
+		{
+			ToTable("Tarefas");
+
+			HasKey(p => p.TarefaId);
+
+			Property(p => p.Nome).IsRequired().HasMaxLength(50);
+			Property(p => p.Descricao).IsOptional().HasMaxLength(250);
+		}
+	}
+}