+++ C#/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.1.0] - 2017-03-27
+### Added
+- CrossCutting (IoC)
+- RepositoryBase e ServiceBase
+### Changed
+- Propriedade Data e Hora da entidade Domain.Entities.Tarefa foram unidos em uma única.
+
+
 ## [0.0.1] - 2017-03-25
 ### Added
 - Estrutura do projeto